chore(docs-ui): Move `.storybook` into docs-ui/
