+++ docs-ui/storybook/main.ts
@@ -5,12 +5,12 @@ import path from 'path';
 
 import {StorybookConfig} from '@storybook/core-common';
 
-import babelConfig from '../babel.config';
+import babelConfig from '../../babel.config';
 
 const toPath = (p: string) => path.join(process.cwd(), p);
 
 const config: StorybookConfig = {
-  stories: ['../docs-ui/stories/**/*.stories.*'],
+  stories: ['../stories/**/*.stories.*'],
   core: {
     builder: 'webpack5',
   },
